+++ clickhouse-data/src/main/java/com/clickhouse/data/ClickHouseDataConfig.java
@@ -78,6 +78,11 @@ public interface ClickHouseDataConfig extends Serializable {
         }
 
         @Override
+        public int getMaxMapperCache() {
+            return config.getMaxMapperCache();
+        }
+
+        @Override
         public int getMaxQueuedBuffers() {
             return config.getMaxQueuedBuffers();
         }
@@ -108,6 +113,11 @@ public interface ClickHouseDataConfig extends Serializable {
         }
 
         @Override
+        public boolean isUseCompilation() {
+            return config.isUseCompilation();
+        }
+
+        @Override
         public boolean isUseObjectsInArray() {
             return config.isUseObjectsInArray();
         }
@@ -128,6 +138,7 @@ public interface ClickHouseDataConfig extends Serializable {
     static final int DEFAULT_READ_BUFFER_SIZE = DEFAULT_BUFFER_SIZE;
     static final int DEFAULT_WRITE_BUFFER_SIZE = DEFAULT_BUFFER_SIZE;
     static final int DEFAULT_MAX_BUFFER_SIZE = 128 * 1024;
+    static final int DEFAULT_MAX_MAPPER_CACHE = 100;
     static final int DEFAULT_MAX_QUEUED_BUFFERS = 512;
     static final int DEFAULT_BUFFER_QUEUE_VARIATION = 100;
 
@@ -138,6 +149,7 @@ public interface ClickHouseDataConfig extends Serializable {
     static final boolean DEFAULT_REUSE_VALUE_WRAPPER = true;
     static final boolean DEFAULT_USE_BINARY_STRING = false;
     static final boolean DEFAULT_USE_BLOCKING_QUEUE = false;
+    static final boolean DEFAULT_USE_COMPILATION = false;
     static final boolean DEFAULT_USE_OBJECT_IN_ARRAY = false;
     static final boolean DEFAULT_WIDEN_UNSIGNED_TYPE = false;
 
@@ -256,6 +268,15 @@ public interface ClickHouseDataConfig extends Serializable {
     }
 
     /**
+     * Gets maximum number of mappers can be cached.
+     *
+     * @return maximum number of mappers can be cached
+     */
+    default int getMaxMapperCache() {
+        return DEFAULT_MAX_MAPPER_CACHE;
+    }
+
+    /**
      * Gets maximum number of buffers can be queued for processing.
      *
      * @return maximum number of buffers can be queued
@@ -326,6 +347,15 @@ public interface ClickHouseDataConfig extends Serializable {
      */
     default boolean isUseBlockingQueue() {
         return DEFAULT_USE_BLOCKING_QUEUE;
+    }
+
+    /**
+     * Checks whether compilation is used in object mapping and serialization.
+     *
+     * @return true if compilation is used; false otherwise
+     */
+    default boolean isUseCompilation() {
+        return DEFAULT_USE_COMPILATION;
     }
 
     /**