+++ clickhouse-data/src/main/java/com/clickhouse/data/ClickHouseRecordMapper.java
@@ -16,28 +16,28 @@ public interface ClickHouseRecordMapper {
      * Wraps iterable records as mapped objects.
      *
      * @param <T>      type of mapped object
+     * @param config   non-null configuration
      * @param columns  non-null list of columns
      * @param records  non-null iterable records
      * @param objClass non-null class of mapped object
+     * @param template optional template object to reuse
      * @return non-null iterable objects
      */
-    static <T> Iterator<T> wrap(List<ClickHouseColumn> columns, Iterator<ClickHouseRecord> records, Class<T> objClass) {
-        return new IterableRecordWrapper<>(columns, records, objClass);
+    static <T> Iterator<T> wrap(ClickHouseDataConfig config, List<ClickHouseColumn> columns,
+            Iterator<ClickHouseRecord> records, Class<T> objClass, T template) {
+        return new IterableRecordWrapper<>(config, columns, records, objClass, template);
     }
 
     /**
      * Gets mapper to turn {@link ClickHouseRecord} into user-defined object.
      *
+     * @param config   non-null configuration
      * @param columns  non-null list of columns
      * @param objClass non-null class of the user-defined object
      * @return non-null user-defined object
      */
-    static ClickHouseRecordMapper of(List<ClickHouseColumn> columns, Class<?> objClass) {
-        if (columns == null || objClass == null) {
-            throw new IllegalArgumentException("Non-null column list and object class are required");
-        }
-
-        return RecordMapperFactory.of(columns, objClass);
+    static ClickHouseRecordMapper of(ClickHouseDataConfig config, List<ClickHouseColumn> columns, Class<?> objClass) {
+        return RecordMapperFactory.of(config, columns, objClass);
     }
 
     /**
@@ -59,7 +59,7 @@ public interface ClickHouseRecordMapper {
      * @param <T>      type of the user-defined object
      * @param r        non-null record
      * @param objClass non-null class of the user-defined object
-     * @param obj      optional object instance to set values
+     * @param obj      reusable object instance to set values
      * @return non-null user-defined object, either new instance of
      *         {@code objClass}, or same as the given {@code obj} when it's not null
      */