+++ clickhouse-data/src/main/java/com/clickhouse/data/mapper/DynamicRecordMapper.java
@@ -11,6 +11,7 @@ import java.util.Map;
 import java.util.function.BiConsumer;
 
 import com.clickhouse.data.ClickHouseColumn;
+import com.clickhouse.data.ClickHouseDataConfig;
 import com.clickhouse.data.ClickHouseDataType;
 import com.clickhouse.data.ClickHouseRecord;
 import com.clickhouse.data.ClickHouseRecordMapper;
@@ -27,7 +28,7 @@ final class DynamicRecordMapper extends AbstractRecordMapper {
                 int modifiers = m.getModifiers();
                 String name = m.getName();
                 if (Modifier.isPublic(modifiers) && !Modifier.isStatic(modifiers) && m.getParameterCount() == 0
-                        && !"asRawObject".equals(name) && name.startsWith("as")) {
+                        && name.startsWith("as")) {
                     Class<?> type = m.getReturnType();
                     if (type != Void.class) {
                         map.put(type.getName(), m);
@@ -57,6 +58,23 @@ final class DynamicRecordMapper extends AbstractRecordMapper {
             } catch (Exception e) {
                 throw new IllegalArgumentException(
                         ClickHouseUtils.format("Failed to set [%s] due to %s", type.getName(), e.getMessage()));
+            }
+        }
+    }
+
+    static class ObjectSetter extends PropertySetter {
+        ObjectSetter(int valueIndex, Method s) {
+            super(valueIndex, Object.class, s);
+        }
+
+        @Override
+        public void accept(Object obj, ClickHouseValue val) {
+            try {
+                // asRawObject() might be impacted by use_binary_string option
+                setter.invoke(obj, val.asObject());
+            } catch (Exception e) {
+                throw new IllegalArgumentException(
+                        ClickHouseUtils.format("Failed to set %s due to %s", val, e.getMessage()));
             }
         }
     }
@@ -99,7 +117,9 @@ final class DynamicRecordMapper extends AbstractRecordMapper {
         for (int i = 0; i < len; i++) {
             PropertyInfo p = properties[i];
             Class<?> type = p.setter.getParameterTypes()[0];
-            if (ClickHouseValue.class.isAssignableFrom(type)) {
+            if (Object.class == type) {
+                list.add(new ObjectSetter(p.index, p.setter));
+            } else if (ClickHouseValue.class.isAssignableFrom(type)) {
                 list.add(new ValueSetter(p.index, type, p.setter));
             } else {
                 list.add(new PropertySetter(p.index,
@@ -113,7 +133,7 @@ final class DynamicRecordMapper extends AbstractRecordMapper {
     }
 
     @Override
-    public ClickHouseRecordMapper get(List<ClickHouseColumn> columns) {
+    public ClickHouseRecordMapper get(ClickHouseDataConfig config, List<ClickHouseColumn> columns) {
         return new DynamicRecordMapper(clazz, columns == null ? Collections.emptyList() : columns, constructor);
     }