+++ clickhouse-data/src/main/java/com/clickhouse/data/mapper/RecordMapperFactory.java
@@ -4,28 +4,17 @@ import java.lang.reflect.Constructor;
 import java.lang.reflect.Method;
 import java.lang.reflect.Modifier;
 import java.util.List;
+import java.util.concurrent.atomic.AtomicReference;
 
 import com.clickhouse.data.ClickHouseCache;
 import com.clickhouse.data.ClickHouseColumn;
+import com.clickhouse.data.ClickHouseDataConfig;
 import com.clickhouse.data.ClickHouseRecord;
 import com.clickhouse.data.ClickHouseRecordMapper;
+import com.clickhouse.data.UnloadableClassLoader;
 
 public final class RecordMapperFactory {
-    private static final ClickHouseCache<Class<?>, WrappedMapper> CACHED_MAPPERS;
-    private static final boolean USE_COMPILED_MAPPER;
-
-    static {
-        CACHED_MAPPERS = ClickHouseCache.create(100, 0, RecordMapperFactory::get);
-
-        Class<?> asmClass = null;
-        try {
-            asmClass = Class.forName("org.objectweb.asm.ClassWriter", true,
-                    RecordMapperFactory.class.getClassLoader());
-        } catch (ClassNotFoundException e) {
-            // ignore
-        }
-        USE_COMPILED_MAPPER = asmClass != null;
-    }
+    private static final AtomicReference<ClickHouseCache<Class<?>, WrappedMapper>> CACHED_MAPPERS = new AtomicReference<>();
 
     static WrappedMapper get(Class<?> objClass) {
         Constructor<?> preferredConsutrctor = null;
@@ -59,16 +48,25 @@ public final class RecordMapperFactory {
             }
         }
 
-        // return USE_COMPILED_MAPPER ? new CompiledRecordMapper(objClass) : new DynamicRecordMapper(objClass);
-        return new DynamicRecordMapper(objClass);
+        return UnloadableClassLoader.HAS_ASM ? new CompiledRecordMapper(objClass) : new DynamicRecordMapper(objClass);
     }
 
-    public static ClickHouseRecordMapper of(List<ClickHouseColumn> columns, Class<?> objClass) {
+    public static ClickHouseRecordMapper of(ClickHouseDataConfig config, List<ClickHouseColumn> columns,
+            Class<?> objClass) {
         if (columns == null || objClass == null) {
             throw new IllegalArgumentException("Non-null column list and object class are required");
         }
 
-        return CACHED_MAPPERS.get(objClass).get(columns);
+        ClickHouseCache<Class<?>, WrappedMapper> cache = CACHED_MAPPERS.get();
+        if (cache == null) {
+            cache = ClickHouseCache.create(config != null ? config.getMaxMapperCache()
+                    : ClickHouseDataConfig.DEFAULT_MAX_MAPPER_CACHE, 0, RecordMapperFactory::get);
+            if (!CACHED_MAPPERS.compareAndSet(null, cache)) {
+                cache = CACHED_MAPPERS.get();
+            }
+        }
+
+        return cache.get(objClass).get(config, columns);
     }
 
     private RecordMapperFactory() {