+++ clickhouse-data/src/main/java/com/clickhouse/data/mapper/WrappedMapper.java
@@ -3,8 +3,9 @@ package com.clickhouse.data.mapper;
 import java.util.List;
 
 import com.clickhouse.data.ClickHouseColumn;
+import com.clickhouse.data.ClickHouseDataConfig;
 import com.clickhouse.data.ClickHouseRecordMapper;
 
 public interface WrappedMapper {
-    ClickHouseRecordMapper get(List<ClickHouseColumn> columns);
+    ClickHouseRecordMapper get(ClickHouseDataConfig config, List<ClickHouseColumn> columns);
 }